requirements.txt
functions-framework
flask
flask-cors
google-adk
google-genai
google-cloud-discoveryengine
gunicorn